add more stuff

+++ WtForecaster/src/components/Places/Comment.js
@@ -1,21 +1,65 @@
 import React, {Component} from 'react';
-import { View, Text, StyleSheet , TouchableWithoutFeedback } from 'react-native';
+import { View, Text, StyleSheet , TouchableWithoutFeedback, Alert } from 'react-native';
 import { AirbnbRating } from 'react-native-ratings';
 import axios from '../../utils/axiosConfig';
+import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
 
 class Comment extends Component {
-  state = {
-    isExpanding: false
-  }
+
 
   deleteComment = (id) => {
-    axios.delete(`/places/remove/comment/${this.props.placeId}/${id}`)
-    .then(result => {
-      this.props.onReload();
-    })
-    .catch(error => {
-      alert('Delete item failed')
-    })
+    Alert.alert(
+      'Confirm Deletion',
+      'Delete this comment?',
+      [
+        {
+          text: 'No',
+          style: 'cancel',
+        },
+        {
+          text: 'Yes', 
+          onPress: () => {
+            axios.delete(`/places/remove/comment/${this.props.placeId}/${id}`)
+            .then(result => {
+              this.props.onReload();
+            })
+            .catch(error => {
+              alert('Delete item failed')
+            })
+          },
+          style: 'destructive'
+        },
+      ],
+    );
+  }
+
+  reportComment = () => {
+    Alert.alert(
+      'Report violation',
+      'Report this comment?',
+      [
+        {
+          text: 'No',
+          style: 'cancel',
+        },
+        {
+          text: 'Yes', 
+          onPress: () => {
+            Alert.alert(
+              'Submitted',
+              'Reported to Admin.',
+              [
+                {
+                  text: 'OK',
+                  style: 'cancel',
+                },
+              ],
+            );
+          },
+          style: 'destructive'
+        },
+      ],
+    );
   }
 
   expandComment = () => {
@@ -27,44 +71,37 @@ class Comment extends Component {
   }
 
   render() {
-    let expandSection = null;
     const comment = this.props.comment;
-    if (this.state.isExpanding) {
-      expandSection = (
-        <TouchableWithoutFeedback onPress={() => this.deleteComment(comment.commentId)}>
-          <Text style={styles.deleteComment}>Delete</Text> 
-        </TouchableWithoutFeedback>
-      )
-    }
     return (
-      <TouchableWithoutFeedback onPress={this.expandComment}>
         <View style={styles.comment}>
             <View style={styles.commentHeader}>
-              <Text style={styles.username}>{comment.email === this.props.email ? `You (${comment.user})` : comment.user}</Text>
-              <AirbnbRating 
-                showRating={false}
-                size={15}
-                defaultRating={comment.rate}
-                isDisabled={true}
-              />
+                <Text style={styles.username}>{comment.email === this.props.email ? `${comment.user}(You)` : comment.user}</Text>
+                <View style={{display: 'flex', flexDirection: 'row',}}>
+                  <AirbnbRating 
+                    showRating={false}
+                    size={15}
+                    defaultRating={comment.rate}
+                    isDisabled={true}
+                  />
+                   {comment.email === this.props.email ? 
+                    <Icon name="delete" size={20} color="#ff5253" onPress={() => this.deleteComment(comment.commentId)}/> :
+                    <Icon name="flag-outline" size={20} color="#263144" onPress={() => this.reportComment()} />
+                    }
+                </View>
             </View>
+            
             <View style={styles.commentContent}>
               <Text>{comment.content}</Text>
             </View>
-            {expandSection}
         </View>
-      </TouchableWithoutFeedback>
-      
     );
   }
 }
 
 const styles = StyleSheet.create({
   comment: {
-    marginHorizontal: 10,
-    borderRadius: 5,
+    borderRadius: 20,
     marginTop: 5,
-    padding: 10
   },
   commentHeader: {
     display: 'flex',
@@ -72,17 +109,18 @@ const styles = StyleSheet.create({
     alignItems: 'center',
     justifyContent: 'space-between',
     marginBottom: 5,
-    marginHorizontal: 10,
+    marginHorizontal: 5,
   },
   username: {
     fontSize: 14,
-    fontWeight: '800'
+    color: '#263144',
+    marginRight: 10
   },
   commentContent: {
     backgroundColor: 'white',
-    marginHorizontal: 10,
-    padding: 5,
-    borderRadius: 5
+    paddingVertical: 5,
+    paddingHorizontal: 10,
+    borderRadius: 20
   },
   deleteComment: {
     color: 'red',
